requirements.txt
# Agentic Browser - Phase 1-A Dependencies

# Browser automation
playwright>=1.49.0

# Web framework + WebSocket
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# LLM integration
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
